requirements.txt
albumentations==1.3.0
einops==0.6.0
lpips==0.1.4
numpy==1.24.2
opencv-python==4.7.0.72
opencv-python-headless==4.7.0.72
Pillow==9.4.0
protobuf==4.22.0
PyYAML==6.0
scikit-image==0.19.3
scikit-learn==1.2.1
scipy==1.10.1
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
torch==1.12.0+cu116
torch-fidelity==0.3.0
torchmetrics==0.11.1
torchvision==0.13.0+cu116
tqdm==4.64.1
visdom==0.1.8.9
